+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 0.7.0 (Jan 9, 2018)
+## 0.7.0 (Jun 29, 2018)
 
 * Add `.onOpen` event channel ([#9])
 * Add `.onSend` event channel ([#8])
+* Add `.onError` event channel ([#4])
 
 ## 0.6.0 (Oct 21, 2017)
 
@@ -22,5 +23,6 @@ are replaced with 4 more atomic hadlers:
 * Added changelog :)
 
 
+[#4]: https://github.com/vitalets/websocket-as-promised/issues/4
 [#8]: https://github.com/vitalets/websocket-as-promised/issues/8
 [#9]: https://github.com/vitalets/websocket-as-promised/issues/9